Move get_max_level to common
