+++ common/utils/season_pass.py
@@ -1,10 +1,10 @@
 from datetime import datetime, timezone
-from typing import Optional
+from typing import Optional, Tuple
 
-from sqlalchemy import select
+from sqlalchemy import select, desc
 from sqlalchemy.orm import joinedload
 
-from common.models.season_pass import SeasonPass
+from common.models.season_pass import SeasonPass, Level
 
 
 def get_current_season(sess, include_exp: bool = False) -> Optional[SeasonPass]:
@@ -13,3 +13,12 @@ def get_current_season(sess, include_exp: bool = False) -> Optional[SeasonPass]:
     if include_exp:
         stmt = stmt.options(joinedload(SeasonPass.exp_list))
     return sess.scalar(stmt)
+
+
+def get_max_level(sess) -> Tuple[Level, int]:
+    """
+    Returns max level of season pass and repeating exp.
+    Last one of level table is not a real level. Just for repeating reward.
+    """
+    m1, m2 = sess.scalars(select(Level).order_by(desc(Level.level)).limit(2)).fetchall()
+    return m2, abs(m1.exp - m2.exp)
