+++ season_pass/utils.py
@@ -19,12 +19,3 @@ def verify_token(authorization: Annotated[str, Header()]):
             raise ExpiredSignatureError()
     except Exception:
         raise HTTPException(status_code=401, detail="Not Authorized")
-
-
-def get_max_level(sess) -> Tuple[Level, int]:
-    """
-    Returns max level of season pass and repeating exp.
-    Last one of level table is not a real level. Just for repeating reward.
-    """
-    m1, m2 = sess.scalars(select(Level).order_by(desc(Level.level)).limit(2)).fetchall()
-    return m2, abs(m1.exp - m2.exp)
